% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{activities}
\alias{activities}
\alias{fac_get_activities}
\alias{fac_get_activities_section}
\alias{fac_get_activities_section_details}
\alias{fac_get_activity_section_details}
\alias{fac_get_activity_section_details_attachments}
\title{Activities}
\usage{
fac_get_activities(unitid, sectionid, userlist,
  activityclassificationfilter, termstart, termfinish,
  includeJournalMetrics = seq(0, 1, 1), showhidden = seq(1, 0, -1),
  data = c("count", "summary", "detailed"), q, limit, offset, ...)

fac_get_activities_section(sectionid, unitid, userlist,
  activityclassificationfilter, activityclassificationpull, termstart,
  termfinish, type, data = c("count", "summary", "detailed"), q, limit,
  offset, ...)

fac_get_activities_section_details(sectionid, unitid, userlist,
  activityclassificationfilter, activityclassificationpull, termstart,
  termfinish, page, last_updated, data = c("count", "summary",
  "detailed"), q, limit, offset, ...)

fac_get_activity_section_details(sectionid, activityid,
  activityclassificationpull, data = c("count", "summary", "detailed"),
  q, limit, offset, ...)

fac_get_activity_section_details_attachments(sectionid, activityid,
  data = c("count", "summary", "detailed"), q, limit, offset, ...)
}
\arguments{
\item{unitid}{integer
retrieve details of a unit with unitid
When defined, activities are limited to users whose primary unit matches the specified unit.
When undefined, all activities are returned.}

\item{sectionid}{integer
Retrieve activities within a sectionid
When defined, limits the activities to only the specified section.
When undefined, all sections are returned.}

\item{userlist}{comma separated vector of string
Values should match the userid of users.
When defined, filter returned data to the specified users.
When undefined, all users are returned.}

\item{activityclassificationfilter}{comma seperated vector of string
When multiple items are specified, queries are constructed using AND.
When defined, activities are filtered to those with the specified choice for the specified activity classification.
When undefined, activities are not filtered by activity classifications.
The activityclassificationidid and activityclassificationchoicevalue are returned by the Activity Classification Details resource in “detailed” output.}

\item{termstart}{string
Value should match the termid of terms.
When defined, limits classification data to the specified starting term.
When undefined, defaults to the current term.}

\item{termfinish}{string
Value should match the termid of terms.
When defined, limits classification data to the specified ending term.
When undefined, defaults to the current term.}

\item{includeJournalMetrics}{integer
Includes Journal Metrics when data=detailed.
Allowed values are: \* 1 - Show \* 0 - (default) Do not show}

\item{showhidden}{integer
Tells the API whether it should respect the 'Publicly Displayed' activity classification in the activity data set that is output.
Allowed values are:
1 - (default) The 'Publicly Displayed' activity classification is ignored and those activities that a faculty member requested hidden ARE shown, or
0 - The 'Publicly Displayed' activity classification is honored and those activities that a faculty member requested hidden ARE NOT shown.}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{activityclassificationpull}{comma separated vector of integer
When defined, returned data includes activity classification values for the specified activity classifications.
When undefined, no activity classification data is included.}

\item{type}{string
Value should match the exact name of a Scholarly Activity and Creative Productions sub-type
When defined, limits activities matched to the specified scholarly activity sub-type
(Note this parameter ONLY applies to records retrieved from sectionid=-21 (i.e. 'Scholarly Activities and Creative Productions'). It is ignored for activities pulled from any other section.}

\item{page}{integer
Combines with the limit parameter and used to pull a given page of activities.}

\item{last_updated}{string
Determines the exact date of when an activity was last updated.
Accepted format is YYYY-MM-DD
When provided,the returned activities are limited to those modified on the provided date}

\item{activityid}{integer
Retrieve an activity with an activityid}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve details of Activities.
}
\note{
API documentation \url{https://faculty180.interfolio.com/swagger/ui/#tag/Activities}
}
\examples{
## get count of all activities
r <- fac_get_activities(unitid = 1)
content(r)

}
