% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_input.R
\name{activity_input}
\alias{activity_input}
\alias{fac_get_activity_input_approval_status}
\alias{fac_get_activity_input_form_name_units}
\title{Activity Input}
\usage{
fac_get_activity_input_approval_status(approval_id, faculty_id, ...)

fac_get_activity_input_form_name_units(formName, ...)
}
\arguments{
\item{approval_id}{integer
Specify the approval process to lookup.}

\item{faculty_id}{integer
Specify the facultyid of the user.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{formName}{string
Specify the name of the form to lookup.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieves activity input
}
\note{
API documentation \url{https://faculty180.interfolio.com/swagger/ui/#tag/Activity-Input}
}
\examples{
## need some examples

}
