% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{courses}
\alias{courses}
\alias{fac_get_course_prefixes}
\alias{fac_get_courses}
\alias{fac_get_courses_taught}
\alias{fac_get_courses_taught_id}
\alias{fac_get_courses_taught_id_attachment}
\title{Courses}
\usage{
fac_get_course_prefixes(unitid, data, q, limit, offset, ...)

fac_get_courses(prefixes, extra = c("title", "description",
  "credithours", "lecturehours", "labhours", "extrahours",
  "extrahourslabel", "courselevel", "cipcode"), data, q, limit, offset,
  ...)

fac_get_courses_taught(termid, userlist, data, q, limit, offset, ...)

fac_get_courses_taught_id(coursestaughtid, termid, userlist, data, q,
  limit, offset, ...)

fac_get_courses_taught_id_attachment(coursestaughtid, data, q, limit,
  offset, ...)
}
\arguments{
\item{unitid}{comma separated vector of integers
When defined, filters for course with given course prefix.
When undefined, all prefixes are returned.}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{prefixes}{comma separated vector of string
When defined, filters for course with given course prefix.
When undefined, all prefixes are returned.}

\item{extra}{comma separated vector of string
When included, additional details are indexed and included in the returned array.}

\item{termid}{string
Value should match the termid of terms.
When defined, filter courses that were taught during the specified term.
When undefined, courses are not filtered by term.}

\item{userlist}{comma separated vector of string
Values should match the userid of users.
When defined, filter returned courses by the specified users.
When undefined, courses are not filtered by users}

\item{coursestaughtid}{string
Specify a coursestaughtid to retrieve details.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve details of Courses (Prefixes, Courses Taught, Attachments).
}
\note{
API documentation \url{https://faculty180.interfolio.com/swagger/ui/#tag/Courses}
}
\examples{
## get count of course prefixes
r <- fac_get_course_prefixes(unitid = 1)
content(r)

}
