% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_attachments.R
\name{download_attachments}
\alias{download_attachments}
\alias{fac_get_download_attachments}
\title{Download Attachments}
\usage{
fac_get_download_attachments(attachmentid, ...)
}
\arguments{
\item{attachmentid}{integer
Retrieves an attachment.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieves an attachment.
}
\note{
API documentation \url{https://faculty180.interfolio.com/swagger/ui/#tag/Download-Attachments}
}
\examples{
## get an attachment ???

}
