% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{evaluations}
\alias{evaluations}
\alias{fac_get_evaluations}
\alias{fac_get_evaluations_search}
\title{Evaluations}
\usage{
fac_get_evaluations(unitid, data = c("count", "summary", "detailed"), q,
  limit, offset, ...)

fac_get_evaluations_search(title, process_id, unit_id, self_evaluation,
  template_flag = seq(0, 2, 1), ...)
}
\arguments{
\item{unitid}{integer
Retrieve an Activity Classification with a specific unitid.
When defined, filter classifications to the specified unit.
When undefined, filters classifications from the highest unit accessible.}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{title}{string
The title assigned to the evaluation}

\item{process_id}{integer
The process id associated with the evaluation process in which the evaluation resides.}

\item{unit_id}{integer
The unit id associated with the unit in which the evaluation resides.}

\item{self_evaluation}{logical
When true, the serach will be limited to self_evaluations.}

\item{template_flag}{integer
Retrieve evaluations that are not templates by using 0. Evaluations that are templates * with a value of 1, or evaluations that are workflow templates with a value of 2.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve details of users' evaluations.
}
\note{
API documentation \url{https://faculty180.interfolio.com/swagger/ui/#tag/Evaluations}
}
\examples{
\dontrun{
## retrieves details for users' evaluations
r <- fac_get_evaluations(data = "detailed")
content(r)

## searches for an existing evaluation
r <- fac_get_evaluations_search()
content(r)
}

}
