% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{fac_get}
\alias{fac_get}
\title{Faculty180 GET}
\usage{
fac_get(request_string, query = list(), ...,
  credentials = fac_credentials(), data = c("count", "summary",
  "detailed"), q, limit, offset)
}
\arguments{
\item{request_string}{string
API endpoint to call}

\item{query}{named list
Passed to \code{\link[httr]{modify_url}} to build query strings.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{credentials}{function
Returns named list of credentials. See \code{\link{fac_credentials}}}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Base function to perform any HMAC GET request.
}
