% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faq.R
\name{faq}
\alias{faq}
\alias{fac_get_faq}
\title{FAQ}
\usage{
fac_get_faq(pending, ...)
}
\arguments{
\item{pending}{logical
When true, only pending FAQ items that are unanswered will be returned.
Only available for * institutional administrators.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieves frequently asked questions.
}
\note{
API documentation \url{https://faculty180.interfolio.com/swagger/ui/#tag/FAQ}
}
\examples{
## get faq
r <- fac_get_faq()
content(r)

}
