% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissions}
\alias{permissions}
\alias{fac_get_users_permissions}
\alias{fac_get_user_permissions}
\title{Permissions}
\usage{
fac_get_users_permissions(userlist, rights = c("1,2", "1", "2"), unitid,
  data, q, limit, offset, ...)

fac_get_user_permissions(userid, rights, unitid, data, q, limit, offset,
  ...)
}
\arguments{
\item{userlist}{string: comma separated vector of users - limits the bulk GET to the users in the list.}

\item{rights}{string - when set to "1" will report only users with Full Admin Rights. When set to "2" will report only users with Report Rights. Defaults to "1,2".}

\item{unitid}{string: comma separated vector of integers - only return permissions for selected units.}

\item{data}{string - determine the extent of data that will be returned.}

\item{q}{string - search term. Use '*' as wildcard.}

\item{limit}{integer - limit query to specified number of records.}

\item{offset}{integer - used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{userid}{string: id of a specific user}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve details of users' permissions.
}
\examples{
## TODO
}
