% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{units}
\alias{units}
\alias{fac_get_units}
\alias{fac_get_unit}
\title{Units}
\usage{
fac_get_units(unitid, data = c("count", "summary", "detailed"), q, limit,
  offset, depth, ...)

fac_get_unit(unitid, data = c("count", "summary", "detailed"), q, limit,
  offset, ...)
}
\arguments{
\item{unitid}{comma separated vector of integers
Retrieve a unit with specific unitid(s).
When defined, the unit specified and any sub-units are returned down the whole depth of the unit tree.
When undefined, highest unit accessible and any sub-units are returned.
Multiple values can be sent as a comma separated list.}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{depth}{integer
When defined, specifies how deep to recurse down the unit tree to find subunits.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Create & retrieve details of academic units.
}
\note{
API documentation: \url{https://faculty180.interfolio.com/swagger/ui/#tag/Units}
}
\examples{
\dontrun{
## retrieves a listing of academic units
r <- fac_get_units(data = "detailed")
content(r)

## retrieves a listing of academic units
r <- fac_get_units(data = "detailed", unitid = 2)
content(r)

## retrieves details on the specific unit
r <- fac_get_unit(unitid = 2, data = "detailed")
content(r)
}

}
