% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_data.R
\name{user_data}
\alias{user_data}
\alias{fac_get_user_data}
\title{User Data}
\usage{
fac_get_user_data(unitid, sectionid, category, termstart, termfinish,
  userlist, q, limit, offset, ...)
}
\arguments{
\item{unitid}{integer
Retrieve a unit with a specific unitid.
When defined, sections are limited to those found at the specified unit.
When undefined, all sections are returned.}

\item{sectionid}{integer
Retrieve a section with a specific sectionid.
When defined, limits the sections to only the specified section.
When undefined, all sections are returned.}

\item{category}{string
Valid values are profile and activity_input.
When defined, limits the sections to the specified form.
When undefined, all sections are returned.}

\item{termstart}{string
Value should match the termid of terms.
When defined, limits activities to the specified starting term.
When undefined, the furthest back term is used.}

\item{termfinish}{string
Value should match the termid of terms.
When defined, limits activities to the specified ending term.
When undefined, defaults to the current term.}

\item{userlist}{comma seperated vector of string
Values should match the userid of users.
When defined, filter returned data to the specified users.
When undefined, all users are returned.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieves a users activity information.
}
\note{
API documentation \url{https://faculty180.interfolio.com/swagger/ui/#tag/User-Data}
}
\examples{
\dontrun{
## retrieves all of the user's activities with section information
r <- fac_get_user_data(userlist = "m123456,m789123")
content(r)

## retrieves all of the user's activities with section information
r <- fac_get_user_data(userlist = "m123456,m789123", category = "activity_input")
content(r)
}

}
