% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitas.R
\name{vitas}
\alias{vitas}
\alias{fac_get_vitas}
\alias{fac_get_vita}
\alias{fac_get_vita_templates}
\title{Vitas}
\usage{
fac_get_vitas(unitid, data = c("count", "summary", "detailed"), q, limit,
  offset, ...)

fac_get_vita(vitaid, userid, termstart, termfinish, output = c("json",
  "html", "html_clean"), q, limit, offset, ...)

fac_get_vita_templates(tenant_id, search_text, limit, page, sort_by,
  descending, pid, ...)
}
\arguments{
\item{unitid}{integer
Enter a unitid to retrieve evaluation data from that specific unit.
When defined, filter evaluations to the specified unit.
When undefined, filters evaluations from the highest unit accessible.}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{vitaid}{integer
Retrieves details for the specified vitae.}

\item{userid}{string
Retrieves details on the specified user(s).}

\item{termstart}{string
Value should match the termid of terms.
When defined, limits classification data to the specified starting term.
When undefined, defaults to the current term.}

\item{termfinish}{string
Value should match the termid of terms.
When defined, limits classification data to the specified ending term.
When undefined, defaults to the current term.}

\item{output}{string
A list of Vitaes.
If output=json
Returns an array of values for the specified vitas, as well as enabled sections on the vita.
If output=html
Directly displays vita requested as an HTML document with Styling as it would display directly from the software.
If output=html_clean
Directly displays vita requested as an HTML document with:
All inline styling from DATA180 default styling is removed.
All styling from a user as supplied from rich text edited text is removed (this will be corrected in a later release to try to honor this styling).
Styling applied to each scholarly work for formatting each scholarly citation with the appropriate citation style is retained.}

\item{tenant_id}{integer
Tenant associated with the request.}

\item{search_text}{string
Search term.}

\item{page}{integer
Limit query to specified page of records.}

\item{sort_by}{string
Sort the results by a desired column.}

\item{descending}{logical
When true, the sort column is sorted in descending order.}

\item{pid}{integer
User pid filter for the given request.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve details of users' vitaes.
}
\note{
API documentation \url{https://faculty180.interfolio.com/swagger/ui/#tag/Vitas}
}
\examples{
## get count of vitas
r <- fac_get_vitas()
content(r)

}
