% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{fac_credentials}
\alias{fac_credentials}
\title{Faculty180 API Credentials}
\usage{
fac_credentials(public_key = Sys.getenv("FAC180_DEV_PK"),
  private_key = Sys.getenv("FAC180_DEV_SK"),
  database_id = Sys.getenv("FAC180_DEV_DB"),
  host = Sys.getenv("FAC180_BASE_URL"))
}
\arguments{
\item{public_key}{string
Public key for HMAC encryption}

\item{private_key}{string
Private key for HMAC encryption}

\item{database_id}{string
Database ID to connect to}

\item{host}{string
API host url}
}
\value{
named list of credentials
}
\description{
Helper function to retrieve HMAC credentials.
}
\examples{
fac_credentials()
}
