% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{fac_get}
\alias{fac_get}
\title{Faculty180 GET}
\usage{
fac_get(request_string, query = list(),
  credentials = fac_credentials())
}
\arguments{
\item{request_string}{string - api endpoint to call}

\item{query}{named list - passed to \code{\link[httr]{modify_url}} to build query strings}

\item{credentials}{function - returns named list of credentials. See \code{\link{fac_credentials}}}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Base function to perform any HMAC GET request.
}
\examples{
response <- fac_get("/units")
httr::content(response)
}
