% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{fac_get}
\alias{fac_get}
\title{Faculty180 GET}
\usage{
fac_get(request_string, query = list(), ...,
  credentials = fac_credentials(), data = c("detailed", "summary",
  "count"), q, limit, offset)
}
\arguments{
\item{request_string}{string - api endpoint to call}

\item{query}{named list - optional - passed to \code{\link[httr]{modify_url}} to build query strings}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{credentials}{function - returns named list of credentials. See \code{\link{fac_credentials}}}

\item{data}{string - optional - determine the extent of data that will be returned.}

\item{q}{string - optional - search term. Use '*' as wildcard.}

\item{limit}{integer - optional - limit query to specified number of records.}

\item{offset}{integer - optional - used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Base function to perform any HMAC GET request.
}
