% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{fac_get_terms}
\alias{fac_get_terms}
\title{Terms}
\usage{
fac_get_terms(yearlist, data = c("count", "summary", "detailed"), q,
  limit, offset, ...)
}
\arguments{
\item{yearlist}{string: comma separated vector of integers - retrieve a term by year.}

\item{data}{string - determine the extent of data that will be returned.}

\item{q}{string - search term. Use '*' as wildcard.}

\item{limit}{integer - limit query to specified number of records.}

\item{offset}{integer - used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve academic terms within your institution.
}
\examples{
## get summary information for one term
r <- fac_get_terms(yearlist = "2019", data = "summary")
content(r)

## get detailed information for two terms
r <- fac_get_terms(yearlist = "2018,2019", data = "detailed")
content(r)
}
