% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty_classifications.R
\name{faculty_classifications}
\alias{faculty_classifications}
\alias{fac_get_faculty_classification_data}
\alias{fac_get_faculty_classifications}
\alias{fac_get_faculty_classification_class}
\title{Faculty Classifications}
\usage{
fac_get_faculty_classification_data(userlist, facultyclassificationlist,
  termstart, termfinish, data = c("count", "summary", "detailed"), q,
  limit, offset, ...)

fac_get_faculty_classifications(unitid, data = c("count", "summary",
  "detailed"), q, limit, offset, ...)

fac_get_faculty_classification_class(facultyclassificationid,
  data = c("count", "summary", "detailed"), q, limit, offset, ...)
}
\arguments{
\item{userlist}{comma separated vector of string
Values should match the userid of users.
When defined, filter returned data to the specified users.
When undefined, all users are returned.}

\item{facultyclassificationlist}{comma separated vector of integer
Values should match the facultyclassificationid of faculty classifications.
When defined, filter returned data to the specified faculty classifications.
When undefined, all faculty classifications are returned.}

\item{termstart}{string
Value should match the termid of terms.
When defined, limits classification data to the specified starting term.
When undefined, defaults to the current term.}

\item{termfinish}{string
Value should match the termid of terms.
When defined, limits classification data to the specified ending term.
When undefined, defaults to the current term.}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{unitid}{integer
Retrieve a faculty classification with a specific unitid.
When defined, filters classifications to the specified unit.
When undefined, filters classifications from the highest unit accessible.}

\item{facultyclassificationid}{integer
Retrieves details regarding a specific faculty classification.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve details of Faculty Classifications.
}
\note{
API documentation: \url{https://faculty180.interfolio.com/swagger/ui/#tag/Faculty-Classifications}
}
\examples{
## get count of faculty classification data
r <- fac_get_faculty_classification_data()
content(r)

## get count of faculty classifications
r <- fac_get_faculty_classifications()
content(r)

## get details on faculty classification class 1
r <- fac_get_faculty_classification_class(facultyclassificationid = 1, data = "detailed")
content(r)

## get details on faculty classification class 1
r <- fac_get_faculty_classification_class(facultyclassificationid = 2, data = "detailed")
content(r)

}
