% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissions}
\alias{permissions}
\alias{fac_get_users_permissions}
\alias{fac_get_user_permissions}
\title{Retrieve details of users' permissions.}
\usage{
fac_get_users_permissions(userlist, rights = c("1,2", "1", "2"), unitid,
  data = c("count", "summary", "detailed"), q, limit, offset, ...)

fac_get_user_permissions(userid, rights, unitid, data = c("count",
  "summary", "detailed"), q, limit, offset, ...)
}
\arguments{
\item{userlist}{comma separated vector of string
Enter user(s) to retrieve permissions information.
When defined, limits the bulk GET to the users in the list.
When undefined, all faculty and their permissions are listed}

\item{rights}{string
Defaults to “1,2” when nothing is sent or the value is invalid.
When set to “1”, the system will only report users with Full Admin Rights.
When set to “2”, the system will only report users with Report Rights.
When set to “1,2” all rights will be returned.}

\item{unitid}{comma separated vector of integer
Retrieve a user with specific unitid(s).
When defined, only returns permissions for selected unitids.
When undefined, all units are selected.
Multiple values can be sent as a comma separated list.}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{userid}{string
Retrieves permission details on the specified user.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve details on user's permissions within your institution. A count of user permissions will be returned if no parameters are specified.
}
\note{
API documentation: \url{https://faculty180.interfolio.com/swagger/ui/#tag/Permissions}
}
\examples{
## get count of users with full admin rights
r <- fac_get_users_permissions(rights = "1")
content(r)

## get permissions for user "000187846"
# need ask interfolio
#r <- fac_get_user_permissions(userid = "000187846")
#content(r)

}
