% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{sections}
\alias{sections}
\alias{fac_get_sections}
\alias{fac_get_section}
\title{Sections}
\usage{
fac_get_sections(unitid, sectionid, data = c("count", "summary",
  "detailed"), q, limit, offset, ...)

fac_get_section(sectionid, data = c("count", "summary", "detailed"), q,
  limit, offset, ...)
}
\arguments{
\item{unitid}{integer: limit sections to a specific unit}

\item{sectionid}{integer: limit sections to a specific section}

\item{data}{string - determine the extent of data that will be returned.}

\item{q}{string - search term. Use '*' as wildcard.}

\item{limit}{integer - limit query to specified number of records.}

\item{offset}{integer - used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve details of Sections.
}
\examples{
git er done

}
