% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{sections}
\alias{sections}
\alias{fac_get_sections}
\alias{fac_get_section}
\title{Sections}
\usage{
fac_get_sections(unitid, sectionid, data = c("count", "summary",
  "detailed"), q, limit, offset, ...)

fac_get_section(sectionid, data = c("count", "summary", "detailed"), q,
  limit, offset, ...)
}
\arguments{
\item{unitid}{integer
Retrieve a section with a specific unitid.
When defined, sections are limited to those found at the specified unit.
When undefined, all sections are returned.}

\item{sectionid}{integer
Retrieve a section with a specific sectionid.
When defined, limits the sections to only the specified section.
When undefined, all sections are returned}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve details of Sections.
}
\note{
API documentation: \url{https://faculty180.interfolio.com/swagger/ui/#tag/Sections}
}
\examples{
## get detailed information on unit 1
r <- fac_get_sections(unitid = 1, data = "detailed")
content(r)

# get detailed information on section 7
r <- fac_get_section(sectionid = 7, data = "detailed")
content(r)

}
