% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{units}
\alias{units}
\alias{fac_get_units}
\alias{fac_get_unit}
\title{Units}
\usage{
fac_get_units(unitid, data = c("count", "summary", "detailed"), q, limit,
  offset, depth, ...)

fac_get_unit(unitid, data = c("count", "summary", "detailed"), q, limit,
  offset, ...)
}
\arguments{
\item{unitid}{string: comma separated vector of integers - retrieve units with specific ids.}

\item{data}{string - optional - determine the extent of data that will be returned.}

\item{q}{string - optional - search term. Use '*' as wildcard.}

\item{limit}{integer - optional - limit query to specified number of records.}

\item{offset}{integer - optional - used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{depth}{integer - how deep to recurse down the unit tree to find subunits}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve academic units within your institution.
}
\examples{
## retrieve a summary listing of the first 3 academic units
response <- fac_get_units(data = "summary", limit = 3)
content(response)

## retrieve detailed data on the first academic unit
response <- fac_get_unit(1, "detailed")
content(response)

}
