% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{units}
\alias{units}
\alias{fac_get_units}
\alias{fac_get_unit}
\title{Units}
\usage{
fac_get_units(unitid, data = c("count", "summary", "detailed"), q, limit,
  offset, depth)

fac_get_unit(unitid, data = c("count", "summary", "detailed"), q, limit,
  offset)
}
\arguments{
\item{unitid}{string: comma separated vector of integers - retrieve units with specific ids.}

\item{data}{string - determine the extent of data that will be returned.}

\item{q}{string - search term. Use '*' as wildcard.}

\item{limit}{integer - limit query to specified number of records.}

\item{offset}{integer - used to offset queries that are segmented into chunks with limit.}

\item{depth}{integer - how deep to recurse down the unit tree to find subunits}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve academic units within your institution.
}
\examples{
## retrieve a summary listing of the first 3 academic units
response <- fac_get_units(data = "summary", limit = 3)
httr::content(response)

## retrieve detailed data on the first academic unit
response <- fac_get_unit(1, "detailed")
httr::content(response)

}
