% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{units}
\alias{units}
\alias{fac_get_units}
\alias{fac_get_unit}
\title{Units}
\usage{
fac_get_units(unitid, data = c("count", "summary", "detailed"), q, limit,
  offset, depth, ...)

fac_get_unit(unitid, data = c("count", "summary", "detailed"), q, limit,
  offset, ...)
}
\arguments{
\item{unitid}{string: comma separated vector of integers - retrieve units with specific ids.}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{depth}{integer - how deep to recurse down the unit tree to find subunits}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve academic units within your institution.
}
\note{
API documentation: \url{https://faculty180.interfolio.com/swagger/ui/#tag/Units}
}
\examples{
## retrieve a summary listing of the first 3 academic units
r <- fac_get_units(data = "summary", limit = 3)
content(r)

## retrieve detailed data on the first academic unit
r <- fac_get_unit(unitid = "1", data = "detailed")
content(r)

}
