% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users}
\alias{users}
\alias{fac_get_users}
\alias{fac_get_user}
\alias{fac_get_users_activity_input_request}
\alias{fac_get_users_current}
\title{Users}
\usage{
fac_get_users(unitid, employmentstatus = c("All", "Full Time",
  "Part Time", "Staff", "Inactive", "All Active"), data = c("count",
  "summary", "detailed"), q, limit, offset, ...)

fac_get_user(userid, extra = c("accounttype", "suffix", "honorific",
  "officebuilding", "officenumber", "phone", "gender", "race",
  "nationality", "homephone", "homestreet1", "homestreet2", "homecity",
  "homestate", "homezipcode", "cellphone", "teachinginterest",
  "researchinterest", "rankname", "biography", "emergencycontact",
  "emergencycontactphone", "pager", "personalstreet1", "personalstreet2",
  "personalcity", "personalstate", "personalzip", "url"),
  data = c("count", "summary", "detailed"), q, limit, offset, ...)

fac_get_users_activity_input_request()

fac_get_users_current()
}
\arguments{
\item{unitid}{comma separated vector of integers
Retrieve a user with specific unitid(s)
When defined, filters users with given unitid as their primary unit.
When undefined, all users for all units are returned.
Multiple values can be sent as a comma separated list.}

\item{employmentstatus}{string
When defined, users are filtered by their employment status.}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{userid}{string
Retrieves full profile details on the specified user(s).
By default the only data that is returned for a given user is basic identifying information, if other user details are needed they should be specified using the “extra” parameter.}

\item{extra}{comma seperated vector of string
Multiple values can be sent as a comma separated list.
When included, additional details are indexed and included in the returned array.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Create, update and retrieve details of users.
}
\note{
API documentation \url{https://faculty180.interfolio.com/swagger/ui/#tag/Users}
}
\examples{
\dontrun{
## retrieves a listing of users and basic identifying information for each
r <- fac_get_users(data = "detailed")
content(r)

## retrieves a listing of users and basic identifying information for each
r <- fac_get_users(data = "detailed", employmentstatus = "Full+Time")
content(r)

## retrieves full profile details on the specified user(s)
r <- fac_get_user(userid = "m123456")
content(r)

## retrieves full profile details on the specified user(s)
r <- fac_get_user(userid = "m123456", data = "detailed", extra = "honorific,phone, acounttype")
content(r)

## retrieves a list of active activity input requests
r <- fac_get_users_activity_input_request()
content(r)

## retrieves a list of tenant IDs and is_institutional_admin flag
r <- fac_get_users_current()
content(r)
}

}
