% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users}
\alias{users}
\alias{fac_get_users}
\alias{fac_get_user}
\title{Users}
\usage{
fac_get_users(unitid, employmentstatus = c("All", "Full Time",
  "Part Time", "Staff", "Inactive", "All Active"), data = c("count",
  "summary", "detailed"), q, limit, offset, ...)

fac_get_user(userid, data = c("count", "summary", "detailed"),
  extra = c("accounttype", "suffix", "honorific", "officebuilding",
  "officenumber", "phone", "gender", "race", "nationality", "homephone",
  "homestreet1", "homestreet2", "homecity", "homestate", "homezipcode",
  "cellphone", "teachinginterest", "researchinterest", "rankname",
  "biography", "emergencycontact", "emergencycontactphone", "pager",
  "personalstreet1", "personalstreet2", "personalcity", "personalstate",
  "personalzip", "url"), q, limit, offset, ...)
}
\arguments{
\item{unitid}{string: comma separated vector of integers - retrieve users in units with specific ids.}

\item{employmentstatus}{string - users are filtered by their employment status.}

\item{data}{string - determine the extent of data that will be returned.}

\item{q}{string - search term. Use '*' as wildcard.}

\item{limit}{integer - limit query to specified number of records.}

\item{offset}{integer - used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{userid}{string: id of a specific user.}

\item{extra}{string: comma seperated vector of string - when included, additional details are indexed and included in the returned array.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Create, update, and retreive details of users.
}
\examples{
## get count of users
r <- fac_get_users()
content(r)

## get details of users in unit 3
r <- fac_get_users(unitid = "3", data = "detailed")
content(r)

## get summary of all full time in units 3 and 4
r <- fac_get_users(unitid = "3,4", employmentstatus = "Full Time", data = "summary")
content(r)

<<<<<<< HEAD
## get summary of user "000187846"
r <- fac_get_user(userid = "000187846", data = "summary")
=======
## get summary of user "000315149"
r <- fac_get_user(userid = "000315149", data = "summary")
>>>>>>> d2ca1ef14f22f4b27ba9059522299629b67558d7
content(r)
}
