% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users}
\alias{users}
\alias{fac_get_users}
\alias{fac_get_user}
\title{Create, update and retrieve details of users.}
\usage{
fac_get_users(unitid, employmentstatus = c("All", "Full Time",
  "Part Time", "Staff", "Inactive", "All Active"), data = c("count",
  "summary", "detailed"), q, limit, offset, ...)

fac_get_user(userid, extra = c("accounttype", "suffix", "honorific",
  "officebuilding", "officenumber", "phone", "gender", "race",
  "nationality", "homephone", "homestreet1", "homestreet2", "homecity",
  "homestate", "homezipcode", "cellphone", "teachinginterest",
  "researchinterest", "rankname", "biography", "emergencycontact",
  "emergencycontactphone", "pager", "personalstreet1", "personalstreet2",
  "personalcity", "personalstate", "personalzip", "url"),
  data = c("count", "summary", "detailed"), q, limit, offset, ...)
}
\arguments{
\item{unitid}{comma separated vector of integers
Retrieve a user with specific unitid(s)
When defined, filters users with given unitid as their primary unit.
When undefined, all users for all units are returned.
Multiple values can be sent as a comma separated list.}

\item{employmentstatus}{string
When defined, users are filtered by their employment status.}

\item{data}{string
Determines the extent of data that will be returned.
Count: Returns count of all activities found.
Summary: Returns an array of sections and counts of activities within that section.
Detailed: Returns an array of sections and a subarray of activityids for each section.}

\item{q}{string
Search term. Use '*' as wildcard.}

\item{limit}{integer
Limit query to specified number of records.}

\item{offset}{integer
Used with limit to return records in batches, e.g., if limit = 20 and offset = 21, records 21-40 will be returned.}

\item{...}{additional arguments passed to \code{\link[httr]{GET}}.}

\item{userid}{string
Retrieves full profile details on the specified user(s).
By default the only data that is returned for a given user is basic identifying information, if other user details are needed they should be specified using the “extra” parameter.}

\item{extra}{comma seperated vector of string
Multiple values can be sent as a comma separated list.
When included, additional details are indexed and included in the returned array.}
}
\value{
a \code{\link[httr]{response}} object
}
\description{
Retrieve details on users within your institution. A count of users will be returned if no parameters are specified.
}
\note{
API documentation \url{https://faculty180.interfolio.com/swagger/ui/#tag/Users}
}
\examples{
## get count of users
r <- fac_get_users()
content(r)

## get details of users in unit 3
r <- fac_get_users(unitid = "3", data = "detailed")
content(r)

## get summary of all full time in units 3 and 4
r <- fac_get_users(unitid = "3,4", employmentstatus = "Full Time", data = "summary")
content(r)

## get summary of user "000187846"
r <- fac_get_user(userid = "000187846", data = "summary")
content(r)

}
